Using Bank factory
Used BankFactory function to get mock list of banks
